item update handler tests
